refactor: read exchange_id and exchange_name from env and log channel subject
