+++ place-orders/place_orders_service.py
@@ -2,53 +2,18 @@ from __future__ import annotations
 
 import asyncio
 import json
-import logging
 import os
 import signal
-from datetime import datetime
-from typing import Literal
 
 import aio_pika
 import ccxt.async_support as ccxt
-import colorlog
 from dotenv import load_dotenv
+from logger_utils import get_logger
+from models import Order
 from pydantic import BaseModel
 from pydantic import Field
 
-logger = logging.getLogger(__name__)
-handler = colorlog.StreamHandler()
-formatter = colorlog.ColoredFormatter(
-    "%(log_color)s%(asctime)s %(levelname)-7s %(message)s%(reset)s",
-    log_colors={
-        "DEBUG": "cyan",
-        "INFO": "cyan",
-        "WARNING": "yellow",
-        "ERROR": "red",
-        "CRITICAL": "red,bg_white",
-    },
-)
-handler.setFormatter(formatter)
-logger.addHandler(handler)
-logger.setLevel(logging.DEBUG)
-
-
-class PriceMessage(BaseModel):
-    symbol: str
-    ask_price: float
-    bid_price: float
-    datetime: datetime
-
-
-class Order(BaseModel):
-    symbol: str
-    quantity: float
-    price: float
-    side: Literal["buy", "sell"]
-    exchange: str
-    order_type: Literal["market", "limit"]
-
-    def key(self) -> str:
-        return f"{self.exchange}:{self.symbol}:{self.side}"
+logger = get_logger(__name__)
 
 
 async def load_exchange(exchange_name: str):
@@ -68,6 +33,7 @@ async def load_exchange(exchange_name: str):
 class Observer(BaseModel):
     channel: aio_pika.Channel | None = None
     exchange: ccxt.Exchange | None = None
+    exchange_name: str
     last_order_ids: dict[str, str] = Field(default_factory=dict)
 
     class Config:
@@ -75,10 +41,24 @@ class Observer(BaseModel):
 
     async def on_order_update(self, message: aio_pika.IncomingMessage):
         try:
+            logger.debug("Received order update from %s", message.routing_key)
             order = Order(**json.loads(message.body))
             if self.exchange is None:
                 return
 
+            if order.order_type == "limit":
+                await self.execute_limit_order(order)
+            elif order.order_type == "market":
+                await self.execute_market_order(order)
+
+        except Exception as e:
+            logger.error(f"Error processing order update: {e}")
+
+        finally:
+            await message.ack()
+
+    async def execute_limit_order(self, order: Order):
+        try:
             last_order_id = self.last_order_ids.get(order.key())
 
             if last_order_id is not None:
@@ -97,31 +77,31 @@ class Observer(BaseModel):
                 type=order.order_type,
             )
             self.last_order_ids[order.key()] = result["id"]
-            logger.debug(f"New order on {order.key()}: {result['id']}")
+            logger.info(f"New order on {order.key()}: {result['id']}")
         except Exception as e:
-            logger.error(f"Error processing order update: {e}")
+            logger.error(f"Error executing limit order: {e} for {order.key()}")
 
-        finally:
-            await message.ack()
-
-    async def on_taker_update(self, message: aio_pika.IncomingMessage):
-        pass
+    async def execute_market_order(self, order: Order):
+        try:
+            result = await self.exchange.create_order(
+                symbol=order.symbol,
+                side=order.side,
+                amount=order.quantity,
+                type=order.order_type,
+            )
+            logger.warning(f"New order on {order.key()}: {result}")
+        except Exception as e:
+            logger.error(f"Error executing market order: {e} for {order.key()}")
 
     async def subscribe(self, connection: aio_pika.Connection):
         self.channel = await connection.channel()
 
         # Declare maker queue
-        queue_name = f"order_updates_{self.exchange_id}"
-        logger.info(f"Declaring queue {queue_name}")
+        queue_name = f"order_updates_{self.exchange_name}"
+
         queue = await self.channel.declare_queue(queue_name)
 
         await queue.consume(self.on_order_update)
-
-        # Declare taker queue
-        queue_name = f"taker_updates_{self.exchange_id}"
-        logger.info(f"Declaring queue {queue_name}")
-        queue = await self.channel.declare_queue(queue_name)
-        await queue.consume(self.on_taker_update)
 
     async def close(self):
         if self.exchange:
@@ -130,9 +110,11 @@ class Observer(BaseModel):
 
 async def main():
     load_dotenv(".env")
-    exchange = await load_exchange("kraken")
+    exchange_id = os.getenv("EXCHANGE_ID")
+    exchange_name = os.getenv("EXCHANGE_NAME")
+    exchange = await load_exchange(exchange_id)
     connection = await aio_pika.connect("amqp://guest:guest@rabbitmq/")
-    observer = Observer(exchange=exchange)
+    observer = Observer(exchange=exchange, exchange_name=exchange_name)
 
     loop = asyncio.get_running_loop()
     main_task = asyncio.Future()
