+++ price-watch/requirements.txt
@@ -56,6 +56,8 @@ pydantic==2.9.2
     # via price-watch (pyproject.toml)
 pydantic-core==2.23.4
     # via pydantic
+python-dotenv==1.0.1
+    # via price-watch (pyproject.toml)
 requests==2.32.3
     # via ccxt
 setuptools==75.2.0
