+++ trade-watch/trades_service.py
@@ -1,64 +1,17 @@
 from __future__ import annotations
 
 import asyncio
-import logging
 import os
-from datetime import datetime
 
 import aio_pika
 import ccxt.pro as ccxt
-import colorlog
 from dotenv import load_dotenv
+from logger_utils import get_logger
+from models import TradeMessage
 from pydantic import BaseModel
 from pydantic import Field
 
-
-logger = logging.getLogger(__name__)
-handler = colorlog.StreamHandler()
-formatter = colorlog.ColoredFormatter(
-    "%(log_color)s%(asctime)s %(levelname)-7s %(message)s%(reset)s",
-    log_colors={
-        "DEBUG": "blue",
-        "INFO": "blue",
-        "WARNING": "yellow",
-        "ERROR": "red",
-        "CRITICAL": "red,bg_white",
-    },
-)
-handler.setFormatter(formatter)
-logger.addHandler(handler)
-logger.setLevel(logging.DEBUG)
-
-
-class TradeMessage(BaseModel):
-    symbol: str
-    amount: float
-    price: float
-    exchange: str
-    side: str
-    datetime: datetime
-
-    def __str__(self) -> str:
-        return f"{self.symbol} {self.side} {self.amount} @ {self.price}"
-
-    @classmethod
-    def from_ccxt(cls, trade: dict, exchange_id: str):
-        try:
-            if trade["timestamp"] is None:
-                dt = datetime.now()
-            else:
-                dt = datetime.fromtimestamp(float(trade["timestamp"]) / 1000)
-            return cls(
-                datetime=dt,
-                symbol=trade["symbol"],
-                amount=trade["amount"],
-                price=trade["price"],
-                exchange=exchange_id,
-                side=trade["side"],
-            )
-        except Exception as e:
-            logger.error(f"Error parsing trade {trade}: {e}")
-            return None
+logger = get_logger(__name__)
 
 
 class Observer(BaseModel):
@@ -70,7 +23,6 @@ class Observer(BaseModel):
                 trades = await exchange.watch_my_trades()
                 for trade in trades:
                     new_message = TradeMessage.from_ccxt(trade, exchange.id)
-                    logger.info(f"Observed from {exchange.id}: {new_message}")
                     yield new_message
             except Exception as e:
                 logger.error(f"Error watching {self.symbol}: {e}")
@@ -82,16 +34,14 @@ class Observer(BaseModel):
 async def producer(observer: Observer, exchange: ccxt.Exchange, rabbitmq_url: str):
     connection = await aio_pika.connect_robust(rabbitmq_url)
     channel = await connection.channel()
-    queue_name = f"trade_updates_{observer.exchange_id}"
-    logger.info(f"Declaring queue {queue_name}")
-    queue = await channel.declare_queue(queue_name)
 
     try:
         async for trade_message in observer.consume(exchange):
+            logger.info("New trade to %s: %s", trade_message.subject(), trade_message)
             message = aio_pika.Message(body=trade_message.model_dump_json().encode())
             await channel.default_exchange.publish(
                 message,
-                routing_key=queue.name,
+                routing_key=trade_message.subject(),
             )
     finally:
         await connection.close()
@@ -114,8 +64,11 @@ async def load_exchange(exchange_name: str):
 
 
 async def main() -> None:
-    exchange = await load_exchange("bitfinex2")
-    observer = Observer(symbol="BTC/USDT", exchange_id="bitfinex")
+    exchange_id = os.getenv("EXCHANGE_ID")
+    exchange_name = os.getenv("EXCHANGE_NAME")
+
+    exchange = await load_exchange(exchange_id)
+    observer = Observer(symbol="BTC/USDT", exchange_id=exchange_name)
     await producer(observer, exchange, "amqp://guest:guest@rabbitmq/")
